remove SDKv2 DS
