+++ internal/service/cluster/data_source_clusters.go
@@ -6,414 +6,183 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/constant"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/service/advancedcluster"
 	matlas "go.mongodb.org/atlas/mongodbatlas"
+
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/types"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/id"
+
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
 )
 
-func DataSourceClusters() *schema.Resource {
-	return &schema.Resource{
-		ReadContext: dataSourceMongoDBAtlasClustersRead,
-		Schema: map[string]*schema.Schema{
-			"project_id": {
-				Type:     schema.TypeString,
+const (
+	clustersDataSourceName = "clusters"
+)
+
+var _ datasource.DataSource = &clustersDS{}
+var _ datasource.DataSourceWithConfigure = &clustersDS{}
+
+func PluralDataSource() datasource.DataSource {
+	return &clustersDS{
+		DSCommon: config.DSCommon{
+			DataSourceName: clustersDataSourceName,
+		},
+	}
+}
+
+type clustersDS struct {
+	config.DSCommon
+}
+
+func (d *clustersDS) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{
+		Attributes: map[string]schema.Attribute{
+			"id": schema.StringAttribute{ // required by hashicorps terraform plugin testing framework: https://github.com/hashicorp/terraform-plugin-testing/issues/84#issuecomment-1480006432
+				DeprecationMessage: "Please use each cluster's id attribute instead",
+				Computed:           true,
+			},
+			"project_id": schema.StringAttribute{
 				Required: true,
 			},
-			"results": {
-				Type:     schema.TypeList,
+			"results": schema.ListNestedAttribute{
 				Computed: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"advanced_configuration": advancedcluster.ClusterAdvancedConfigurationSchemaComputed(),
-						"auto_scaling_disk_gb_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"auto_scaling_compute_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"auto_scaling_compute_scale_down_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"backup_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"bi_connector_config": {
-							Type:     schema.TypeList,
-							Computed: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"enabled": {
-										Type:     schema.TypeBool,
-										Computed: true,
-									},
-									"read_preference": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-								},
-							},
-						},
-						"cluster_type": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"connection_strings": {
-							Type:     schema.TypeList,
-							Computed: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"standard": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"standard_srv": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"aws_private_link": {
-										Type:     schema.TypeMap,
-										Computed: true,
-									},
-									"aws_private_link_srv": {
-										Type:     schema.TypeMap,
-										Computed: true,
-									},
-									"private": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"private_srv": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"private_endpoint": {
-										Type:     schema.TypeList,
-										Computed: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"connection_string": {
-													Type:     schema.TypeString,
-													Computed: true,
-												},
-												"endpoints": {
-													Type:     schema.TypeList,
-													Computed: true,
-													Elem: &schema.Resource{
-														Schema: map[string]*schema.Schema{
-															"endpoint_id": {
-																Type:     schema.TypeString,
-																Computed: true,
-															},
-															"provider_name": {
-																Type:     schema.TypeString,
-																Computed: true,
-															},
-															"region": {
-																Type:     schema.TypeString,
-																Computed: true,
-															},
-														},
-													},
-												},
-												"srv_connection_string": {
-													Type:     schema.TypeString,
-													Computed: true,
-												},
-												"srv_shard_optimized_connection_string": {
-													Type:     schema.TypeString,
-													Computed: true,
-												},
-												"type": {
-													Type:     schema.TypeString,
-													Computed: true,
-												},
-											},
-										},
-									},
-								},
-							},
-						},
-						"disk_size_gb": {
-							Type:     schema.TypeFloat,
-							Computed: true,
-						},
-						"encryption_at_rest_provider": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"mongo_db_major_version": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"num_shards": {
-							Type:     schema.TypeInt,
-							Computed: true,
-						},
-						"provider_backup_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"provider_instance_size_name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"provider_name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"backing_provider_name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"provider_disk_iops": {
-							Type:     schema.TypeInt,
-							Computed: true,
-						},
-						"provider_disk_type_name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"provider_encrypt_ebs_volume": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"provider_region_name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"provider_volume_type": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"provider_auto_scaling_compute_min_instance_size": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"provider_auto_scaling_compute_max_instance_size": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"replication_factor": {
-							Type:     schema.TypeInt,
-							Computed: true,
-						},
-						"replication_specs": {
-							Type:     schema.TypeList,
-							Computed: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"id": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"num_shards": {
-										Type:     schema.TypeInt,
-										Computed: true,
-									},
-									"regions_config": {
-										Type:     schema.TypeSet,
-										Computed: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"region_name": {
-													Type:     schema.TypeString,
-													Computed: true,
-												},
-												"electable_nodes": {
-													Type:     schema.TypeInt,
-													Computed: true,
-												},
-												"priority": {
-													Type:     schema.TypeInt,
-													Computed: true,
-												},
-												"read_only_nodes": {
-													Type:     schema.TypeInt,
-													Computed: true,
-												},
-												"analytics_nodes": {
-													Type:     schema.TypeInt,
-													Computed: true,
-												},
-											},
-										},
-									},
-									"zone_name": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-								},
-							},
-						},
-						"mongo_db_version": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"mongo_uri": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"mongo_uri_updated": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"mongo_uri_with_options": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"pit_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"paused": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-						"srv_address": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"state_name": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"labels": {
-							Type:       schema.TypeSet,
-							Computed:   true,
-							Deprecated: fmt.Sprintf(constant.DeprecationParamByDateWithReplacement, "September 2024", "tags"),
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"key": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-									"value": {
-										Type:     schema.TypeString,
-										Computed: true,
-									},
-								},
-							},
-						},
-						"tags":                   &advancedcluster.DSTagsSchema,
-						"snapshot_backup_policy": computedCloudProviderSnapshotBackupPolicySchema(),
-						"container_id": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"version_release_system": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"termination_protection_enabled": {
-							Type:     schema.TypeBool,
-							Computed: true,
-						},
-					},
+				NestedObject: schema.NestedAttributeObject{
+					Attributes: clusterDSAttributes(),
 				},
 			},
 		},
 	}
 }
 
-func dataSourceMongoDBAtlasClustersRead(ctx context.Context, d *schema.ResourceData, meta any) diag.Diagnostics {
-	// Get client connection.
-	conn := meta.(*config.MongoDBClient).Atlas
-	projectID := d.Get("project_id").(string)
-	d.SetId(id.UniqueId())
+func (d *clustersDS) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+	conn := d.Client.Atlas
+	var clustersConfig tfClustersDSModel
 
-	clusters, resp, err := conn.Clusters.List(ctx, projectID, nil)
+	resp.Diagnostics.Append(req.Config.Get(ctx, &clustersConfig)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	projectID := clustersConfig.ProjectID.ValueString()
+
+	clusters, response, err := conn.Clusters.List(ctx, projectID, nil)
 	if err != nil {
-		if resp != nil && resp.StatusCode == http.StatusNotFound {
-			return nil
+		if response != nil && response.StatusCode == http.StatusNotFound {
+			resp.Diagnostics.AddError("clusters not found in Atlas", fmt.Sprintf("error reading cluster list for project(%s): %s", projectID, err))
+			return
 		}
-
-		return diag.FromErr(fmt.Errorf("error reading cluster list for project(%s): %s", projectID, err))
+		resp.Diagnostics.AddError("error in getting clusters from Atlas", fmt.Sprintf("error reading cluster list for project(%s): %s", projectID, err))
+		return
 	}
 
-	if err := d.Set("results", flattenClusters(ctx, d, conn, clusters)); err != nil {
-		return diag.FromErr(fmt.Errorf(advancedcluster.ErrorClusterSetting, "results", d.Id(), err))
+	newClustersState, err := newTFClustersDSModel(ctx, conn, clusters, projectID)
+	if err != nil {
+		resp.Diagnostics.AddError("error while getting clusters results from Atlas", fmt.Sprintf("error reading cluster list for project(%s): %s", projectID, err))
+		return
 	}
 
-	return nil
+	resp.Diagnostics.Append(resp.State.Set(ctx, &newClustersState)...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	newClustersState.ID = types.StringValue(id.UniqueId())
 }
 
-func flattenClusters(ctx context.Context, d *schema.ResourceData, conn *matlas.Client, clusters []matlas.Cluster) []map[string]any {
-	results := make([]map[string]any, 0)
-
-	for i := range clusters {
-		snapshotBackupPolicy, err := flattenCloudProviderSnapshotBackupPolicy(ctx, d, conn, clusters[i].GroupID, clusters[i].Name)
-		if err != nil {
-			log.Printf("[WARN] Error setting `snapshot_backup_policy` for the cluster(%s): %s", clusters[i].ID, err)
-		}
-
-		processArgs, _, err := conn.Clusters.GetProcessArgs(ctx, clusters[i].GroupID, clusters[i].Name)
-		log.Printf("[WARN] Error setting `advanced_configuration` for the cluster(%s): %s", clusters[i].ID, err)
-
-		var containerID string
-		if clusters[i].ProviderSettings != nil && clusters[i].ProviderSettings.ProviderName != "TENANT" {
-			containers, _, err := conn.Containers.List(ctx, clusters[i].GroupID,
-				&matlas.ContainersListOptions{ProviderName: clusters[i].ProviderSettings.ProviderName})
-			if err != nil {
-				log.Printf(errorClusterRead, clusters[i].Name, err)
-			}
-
-			containerID = getContainerID(containers, &clusters[i])
-		}
-		result := map[string]any{
-			"advanced_configuration":                  advancedcluster.FlattenProcessArgs(processArgs),
-			"auto_scaling_compute_enabled":            clusters[i].AutoScaling.Compute.Enabled,
-			"auto_scaling_compute_scale_down_enabled": clusters[i].AutoScaling.Compute.ScaleDownEnabled,
-			"auto_scaling_disk_gb_enabled":            clusters[i].BackupEnabled,
-			"backup_enabled":                          clusters[i].BackupEnabled,
-			"provider_backup_enabled":                 clusters[i].ProviderBackupEnabled,
-			"cluster_type":                            clusters[i].ClusterType,
-			"connection_strings":                      advancedcluster.FlattenConnectionStrings(clusters[i].ConnectionStrings),
-			"disk_size_gb":                            clusters[i].DiskSizeGB,
-			"encryption_at_rest_provider":             clusters[i].EncryptionAtRestProvider,
-			"mongo_db_major_version":                  clusters[i].MongoDBMajorVersion,
-			"name":                                    clusters[i].Name,
-			"num_shards":                              clusters[i].NumShards,
-			"mongo_db_version":                        clusters[i].MongoDBVersion,
-			"mongo_uri":                               clusters[i].MongoURI,
-			"mongo_uri_updated":                       clusters[i].MongoURIUpdated,
-			"mongo_uri_with_options":                  clusters[i].MongoURIWithOptions,
-			"pit_enabled":                             clusters[i].PitEnabled,
-			"paused":                                  clusters[i].Paused,
-			"srv_address":                             clusters[i].SrvAddress,
-			"state_name":                              clusters[i].StateName,
-			"replication_factor":                      clusters[i].ReplicationFactor,
-			"provider_auto_scaling_compute_min_instance_size": clusters[i].ProviderSettings.AutoScaling.Compute.MinInstanceSize,
-			"provider_auto_scaling_compute_max_instance_size": clusters[i].ProviderSettings.AutoScaling.Compute.MaxInstanceSize,
-			"backing_provider_name":                           clusters[i].ProviderSettings.BackingProviderName,
-			"provider_disk_iops":                              clusters[i].ProviderSettings.DiskIOPS,
-			"provider_disk_type_name":                         clusters[i].ProviderSettings.DiskTypeName,
-			"provider_encrypt_ebs_volume":                     clusters[i].ProviderSettings.EncryptEBSVolume,
-			"provider_instance_size_name":                     clusters[i].ProviderSettings.InstanceSizeName,
-			"provider_name":                                   clusters[i].ProviderSettings.ProviderName,
-			"provider_region_name":                            clusters[i].ProviderSettings.RegionName,
-			"bi_connector_config":                             advancedcluster.FlattenBiConnectorConfig(clusters[i].BiConnector),
-			"replication_specs":                               flattenReplicationSpecs(clusters[i].ReplicationSpecs),
-			"labels":                                          advancedcluster.FlattenLabels(clusters[i].Labels),
-			"tags":                                            advancedcluster.FlattenTags(clusters[i].Tags),
-			"snapshot_backup_policy":                          snapshotBackupPolicy,
-			"termination_protection_enabled":                  clusters[i].TerminationProtectionEnabled,
-			"version_release_system":                          clusters[i].VersionReleaseSystem,
-			"container_id":                                    containerID,
-		}
-		results = append(results, result)
+func newTFClustersDSModel(ctx context.Context, conn *matlas.Client, clusters []matlas.Cluster, projectID string) (tfClustersDSModel, error) {
+	tfClustersModel := tfClustersDSModel{
+		ID:        types.StringValue(id.UniqueId()),
+		ProjectID: types.StringValue(projectID),
 	}
 
-	return results
+	res, err := newTFClustersDSModelResults(ctx, conn, clusters)
+	if err != nil {
+		return tfClustersModel, fmt.Errorf("error while getting clusters results from Atlas")
+	}
+
+	tfClustersModel.Results = res
+	return tfClustersModel, nil
+}
+
+func newTFClustersDSModelResults(ctx context.Context, conn *matlas.Client, clusters []matlas.Cluster) ([]*tfClusterDSModel, error) {
+	results := make([]*tfClusterDSModel, len(clusters))
+
+	for i := range clusters {
+		cluster := clusters[i]
+
+		snapshotBackupPolicy, err := newTFSnapshotBackupPolicyModelFromAtlas(ctx, conn, cluster.GroupID, cluster.Name)
+		if err != nil {
+			return nil, err
+		}
+
+		advancedConfiguration, err := newTFAdvancedConfigurationModelDSFromAtlas(ctx, conn, cluster.GroupID, cluster.Name)
+		if err != nil {
+			return nil, err
+		}
+
+		var containerID string
+		if cluster.ProviderSettings != nil && cluster.ProviderSettings.ProviderName != "TENANT" {
+			containers, _, err := conn.Containers.List(ctx, cluster.GroupID,
+				&matlas.ContainersListOptions{ProviderName: cluster.ProviderSettings.ProviderName})
+			if err != nil {
+				log.Printf(errorClusterRead, cluster.Name, err)
+			}
+
+			containerID = getContainerID(containers, &cluster)
+		}
+		result := &tfClusterDSModel{
+			AdvancedConfiguration:              advancedConfiguration,
+			AutoScalingComputeEnabled:          types.BoolPointerValue(cluster.AutoScaling.Compute.Enabled),
+			AutoScalingComputeScaleDownEnabled: types.BoolPointerValue(cluster.AutoScaling.Compute.ScaleDownEnabled),
+			AutoScalingDiskGbEnabled:           types.BoolPointerValue(cluster.AutoScaling.DiskGBEnabled),
+			BackupEnabled:                      types.BoolPointerValue(cluster.BackupEnabled),
+			ProviderBackupEnabled:              types.BoolPointerValue(cluster.ProviderBackupEnabled),
+			ClusterType:                        types.StringValue(cluster.ClusterType),
+			ConnectionStrings:                  newTFConnectionStringsModelDS(ctx, cluster.ConnectionStrings),
+			DiskSizeGb:                         types.Float64PointerValue(cluster.DiskSizeGB),
+			EncryptionAtRestProvider:           types.StringValue(cluster.EncryptionAtRestProvider),
+			MongoDBMajorVersion:                types.StringValue(cluster.MongoDBMajorVersion),
+			Name:                               types.StringValue(cluster.Name),
+			NumShards:                          types.Int64PointerValue(cluster.NumShards),
+			MongoDBVersion:                     types.StringValue(cluster.MongoDBVersion),
+			MongoURI:                           types.StringValue(cluster.MongoURI),
+			MongoURIUpdated:                    types.StringValue(cluster.MongoURIUpdated),
+			MongoURIWithOptions:                types.StringValue(cluster.MongoURIWithOptions),
+			PitEnabled:                         types.BoolPointerValue(cluster.PitEnabled),
+			Paused:                             types.BoolPointerValue(cluster.Paused),
+			SrvAddress:                         types.StringValue(cluster.SrvAddress),
+			StateName:                          types.StringValue(cluster.StateName),
+			ReplicationFactor:                  types.Int64PointerValue(cluster.ReplicationFactor),
+
+			ProviderAutoScalingComputeMinInstanceSize: types.StringValue(cluster.ProviderSettings.AutoScaling.Compute.MinInstanceSize),
+			ProviderAutoScalingComputeMaxInstanceSize: types.StringValue(cluster.ProviderSettings.AutoScaling.Compute.MaxInstanceSize),
+			BackingProviderName:                       types.StringValue(cluster.ProviderSettings.BackingProviderName),
+			ProviderDiskIops:                          types.Int64PointerValue(cluster.ProviderSettings.DiskIOPS),
+			ProviderDiskTypeName:                      types.StringValue(cluster.ProviderSettings.DiskTypeName),
+			ProviderEncryptEbsVolume:                  types.BoolPointerValue(cluster.ProviderSettings.EncryptEBSVolume),
+			ProviderInstanceSizeName:                  types.StringValue(cluster.ProviderSettings.InstanceSizeName),
+			ProviderName:                              types.StringValue(cluster.ProviderSettings.ProviderName),
+			ProviderRegionName:                        types.StringValue(cluster.ProviderSettings.RegionName),
+
+			BiConnectorConfig:            newTFBiConnectorConfigModel(cluster.BiConnector),
+			ReplicationSpecs:             newTFReplicationSpecsModel(cluster.ReplicationSpecs),
+			Labels:                       newTFLabelsModel(cluster.Labels),
+			Tags:                         newTFTagsModel(cluster.Tags),
+			SnapshotBackupPolicy:         snapshotBackupPolicy,
+			TerminationProtectionEnabled: types.BoolPointerValue(cluster.TerminationProtectionEnabled),
+			VersionReleaseSystem:         types.StringValue(cluster.VersionReleaseSystem),
+			ContainerID:                  types.StringValue(containerID),
+			ProjectID:                    types.StringValue(cluster.GroupID),
+			ID:                           types.StringValue(cluster.ID),
+		}
+		results[i] = result
+	}
+	return results, nil
+}
+
+type tfClustersDSModel struct {
+	ID        types.String        `tfsdk:"id"`
+	ProjectID types.String        `tfsdk:"project_id"`
+	Results   []*tfClusterDSModel `tfsdk:"results"`
 }
